requirements.txt
# Core libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
scikit-learn==1.3.0

# Deep learning frameworks
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Image processing
opencv-python==4.8.0.76
pillow==10.0.0
albumentations==1.3.1
scikit-image==0.21.0

# Medical imaging
nibabel==5.1.0
pydicom==2.4.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
ipywidgets==8.0.7

# Data loading and management
kagglehub==0.2.5
h5py==3.9.0
pyarrow==12.0.1

# Jupyter and development
jupyter==1.0.0
notebook==6.5.4
ipython==8.15.0
tqdm==4.65.0

# Model training and evaluation
tensorboard==2.13.0
mlflow==2.6.0
optuna==3.3.0
wandb==0.15.8

# State-of-the-art models
transformers==4.31.0
timm==0.9.2  # PyTorch Image Models
segmentation-models==1.0.1
segmentation-models-pytorch==0.3.3

# Utilities
pyyaml==6.0.1
jsonschema==4.19.0
joblib==1.3.2
gdown==4.7.1  # For downloading large files from Google Drive

# Web and API (for potential deployment)
flask==2.3.3
fastapi==0.103.1
uvicorn==0.23.2